completed refactor read-only-object-factory.ts to read-only.model.ts and spec file
