+++ core/templates/domain/topic_viewer/read-only-topic.model.ts
@@ -13,11 +13,8 @@
 // limitations under the License.
 
 /**
- * @fileoverview Factory for creating instances of ReadOnlyTopic from
- * topic data.
+ * @fileoverview Model class for creating new frontend instances of ParamMetadata
  */
-
-import {Injectable} from '@angular/core';
 
 import {ShortSkillSummary} from 'domain/skill/short-skill-summary.model';
 import {
@@ -147,15 +144,7 @@ export class ReadOnlyTopic {
   getClassroomName(): string | null {
     return this._classroomName;
   }
-}
-
-@Injectable({
-  providedIn: 'root',
-})
-export class ReadOnlyTopicObjectFactory {
-  constructor() {}
-
-  createFromBackendDict(
+  static createFromBackendDict(
     topicDataDict: ReadOnlyTopicBackendDict
   ): ReadOnlyTopic {
     let subtopics = topicDataDict.subtopics.map(subtopic => {
