+++ core/templates/services/UpgradedServices.ts
@@ -239,7 +239,7 @@ import {
   // eslint-disable-next-line max-len
 } from 'interactions/RatioExpressionInput/directives/ratio-expression-input-validation.service';
 import {ReadOnlyCollectionBackendApiService} from 'domain/collection/read-only-collection-backend-api.service';
-import {ReadOnlyTopicObjectFactory} from 'domain/topic_viewer/read-only-topic-object.factory';
+import {ReadOnlyTopicObjectFactory} from 'domain/topic_viewer/read-only-topic.model';
 import {ReviewTestBackendApiService} from 'domain/review_test/review-test-backend-api.service';
 import {ReviewTestEngineService} from 'pages/review-test-page/review-test-engine.service';
 import {SchemaDefaultValueService} from 'services/schema-default-value.service';
